hoan thanh bao cao

+++ src/main/java/dao/UsersImpl.java
@@ -3,19 +3,31 @@ package dao;
 import entity.Users;
 import jakarta.persistence.EntityManager;
 import util.XJPA;
-
 import java.util.List;
 
 public class UsersImpl implements UsersDAO {
+
+    @Override
     public Users findById(String id) {
-        return null;
+        EntityManager em = XJPA.getEntityManager();
+        try {
+            return em.find(Users.class, id);
+        } finally {
+            em.close();
+        }
     }
 
+    @Override
     public List<Users> findAll() {
         EntityManager em = XJPA.getEntityManager();
-        return em.createQuery("SELECT u FROM Users u").getResultList();
+        try {
+            return em.createQuery("SELECT u FROM Users u", Users.class).getResultList();
+        } finally {
+            em.close();
+        }
     }
 
+    @Override
     public void create(Users user) {
         EntityManager em = XJPA.getEntityManager();
         try {
@@ -23,10 +35,14 @@ public class UsersImpl implements UsersDAO {
             em.persist(user);
             em.getTransaction().commit();
         } catch (Exception e) {
-            System.out.println(e);
+            em.getTransaction().rollback();
+            throw e;
+        } finally {
+            em.close();
         }
     }
 
+    @Override
     public void update(Users user) {
         EntityManager em = XJPA.getEntityManager();
         try {
@@ -34,18 +50,28 @@ public class UsersImpl implements UsersDAO {
             em.merge(user);
             em.getTransaction().commit();
         } catch (Exception e) {
-            System.out.println(e);
+            em.getTransaction().rollback();
+            throw e;
+        } finally {
+            em.close();
         }
     }
 
-    public void delete(Users user) {
+    @Override
+    public void delete(String id) {
         EntityManager em = XJPA.getEntityManager();
         try {
             em.getTransaction().begin();
-            em.remove(user);
+            Users user = em.find(Users.class, id);
+            if (user != null) {
+                em.remove(user);
+            }
             em.getTransaction().commit();
         } catch (Exception e) {
-            System.out.println(e);
+            em.getTransaction().rollback();
+            throw e;
+        } finally {
+            em.close();
         }
     }
 }
